Update patch set 1
Patch Set 1:

I'm tempted to agree with Joshua, however, it seems like it would be more useful in reporting to have times, even if they don't mean exactly the same thing.  Basically, a RETRY_LIMIT with no times doesn't tell us useful information (it doesn't tell us when things were broken or if they are now fixed).  It also makes it hard to expire old records from the db.

So I think I'm inclined to say that we should always have times.  This patch provides one solution, though another might be to add start/end times in the scheduler or pipeline manager for the kinds of runs which miss it.  That may be preferable since it's more generalized, and we don't have to solve the problem in multiple drivers.

Patch-set: 1
Label: Verified=0