Update patch set 1
Patch Set 1: Code-Review-1

I worry about this -- it will cause actual work to be submitted to mergers, which may not be what people are expecting.  That's why we removed the config validator.  Let's discuss what the overall experience for users would be with a change like this.

Patch-set: 1
Label: Code-Review=-1