Remove project.unparsed_config
This holds the entire configuration for a project as a cache so we
don't need to reload everything on a config change, but so does
project.unparsed_branch_config.  The latter just does it in smaller
chunks.

We started with only unparsed_config, then added unparsed_branch_config
but never fully replaced unparsed_config.  That allowed for a small
gain of speed on rebuilding a configuration at a relatively larger
cost in memory.

Instead, only keep one copy of the cached config, and rebuild it
from individual branches when required.

Change-Id: I7b13c3fa288a19026776069c5e19f0b098b0b317