Update patch set 1
Patch Set 1:

After we upgraded to Gerrit 2.14, we found a few code paths that misbehave. One is when looking for change dependsOn / neededBy, the JSON has a list of patchsets but the change and patchset number is extracted from the ref spec  ref/changes/45/12345 which yields strings.  That cause an equality mismatch when comparing with an int.

The idea Tyler Cipriani had is to have the Change object to normalise its properties 'number' and 'patchset' to always be strings https://gerrit.wikimedia.org/r/#/c/411466/1/zuul/model.py

So I guess we will patch our Zuul 2.6 version with Tyler patch and have them cast to str then work on upstreaming that for Zuul v3 and hopefully figure out a few unit test to cover the feature :-}

Happy week-end!

Patch-set: 1