Use a status code to detect unknown vs. missing tenant
Re-use http status codes for this, since it's just arbitrary
machine-reable info for now anyway. Use 404, since that one's well know
and what we'll be sending on to the user anyway. Use 204 for the tenant
not being ready yet, since it's a valid request there just isn't any
data right now.

Change-Id: Ife3de557d48a12c85671643877106479e77d7fd6