Update patch set 2
Patch Set 2:

My hesitation on this comes from two places:

1) The authors of YAML have expressed an opinion that they would prefer folks use the ".yaml" extension when possible[1].  Of course, that's not binding, but it is, I think, something we should consider.  There are zero platforms where Zuul will run and ".yaml" will not work, so it's certainly possible for us to support only ".yaml" with no loss in functionality.

2) If we support more extensions, it makes Zuul itself more complex (that's easily dealt with -- this change isn't that big).  But it makes it more complex for others (humans or programs) to deal with Zuul configuration.  This feels like the sort of thing you would add for backwards compatability, but we don't need it -- we've never supported ".yml".

I recognize the close relationship Zuul and Ansible have, but it's Ansible has made some peculiar choices in their handling of yaml, and we are not required to repeat those.

Having said all that, I generally subscribe to Postel's law, and I think this change mostly adheres to that (other than it facilitates the issue raised in #2).  So if we really think it's worth it after considering all that, I'm okay adding it.

An alternative middle ground might be to add a warning message to the report that the ".yml" extension is not supported if there's a zuul.yml file in a change.

[1] http://www.yaml.org/faq.html

Patch-set: 2