Start log receiver in executor
The persistent ssh connections mean that we can only forward a
log receiving socket once, yet the current system starts a new
receiver for each task.  Instead, start a single reciever for a
job in the executor, and forward the socket for that receiver
over every ssh connection.

The log receiver is started as a subprocess of the executor,
using multiprocessing.  We don't share any state between the two
processes, so we should avoid issues related to the unreliable
multiprocessing.Manager.  It's merely used to aid in process
management.

Change-Id: I7c0fe477e0224cf85b63a0b1b316244ba8b64b83