Update patch set 1
Patch Set 1: Code-Review-1

Does this overrides site variables?  We don't want anything to be able to override those, because we want a site operator to be able to set a value used by a base playbook and know that no one can override it.  So to merge this we need:

* To resolve the conflict with site-variables somehow.
* Update the precedence documentation here: https://docs.openstack.org/infra/zuul/user/jobs.html#variables
* Add extra vars to the check-vars job in test_playbook to assert the correct precedence.

Patch-set: 1
Label: Code-Review=-1