Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

I'm cautious of the security implications of this change. I'm not sure if this has been discussed and I've missed it elsewhere?

-1: This needs documentation

Should we also consider that if a file ends with .yaml we should attempt to load it with safe_load ignoring file permissions?

Patch-set: 10
Label: Code-Review=-1