Update patch set 14
Patch Set 14: Code-Review-1

Don't hate me ... I think we need to start splitting these up into a change that adds the REST route in zuul-web itself, so we can land that and let it roll live - and then the one adds the dashboard piece.

Reason is that the html/js stuff goes live immediately - while the zuul-web changes require a zuul-web restart. Thus if we land the html changes before the api pieces are live, we'll be rolling broken pages live.

Patch-set: 14
Label: Code-Review=-1