Shift html templates into components
In the current setup, loading each page flashes an unexpanded template
for a brief second until the javascript loads, then the template is
expanded and the page works.

Instead of including the template in the html that is served, make each
page have a component with a template. Use the webpack html-loader to
pull these templates in as strings during the compile phase and pass
them to the component template argument. This pattern will become more
beneficial when we start refactoring status, as various elements of the
display can become components themselves.

Change the extensions to html. For the angular component templates,
angular is expanding the chunks.

As a result of doing this, we can extract the boilerplate html into a
single shared template that webpack will expand during compilation into
each individual file. Move that template file to web/main.ejs so that
it's clear that it's being templated by webpack as opposed to the html
chunks in templates that are templated by angular.

In the controller javascript code, shift from using $scope (which
sets variables into the global scope) to using this - which adds things
to the controller object. This keeps code more locally scoped. Also
update a few of the for loops to use the for/of syntax from ES6.

Change-Id: Iaf2b5296618e4f689d27c1a0308aae024f6828cc