Serve the static files more dynamically
This is really just an excuse to write that commit subject.

In the current code, we do a scan of static_dir at startup and register
static routes for each of the files we find there. This lets the
webpack-built content be put into zuul/web/static and for things to
work.

Unfortunately for that plane, when new html/js is rolled out into the
static dir for a running zuul-web we don't have routes for the files.

Instead of doing that, define fallthrough routes that are in the routing
table after the other things that will serve the static files. This way
GET /openstack/projects will get handled by the REST API handler that is
defined, but GET /openstack/projects.html will get picked up by the
static handler.

aiohttp has a static directory handler, that could be used - EXCEPT it
can't be used with the '/{tenant}/{path:.*}' route.

Change-Id: I9e2956270982c75c2c7d31f4982549f0898e1ae9