Use arrow functions for http callbacks
We currently have to do a trick of saving the this variable into another
variable so that we can capture it in a callback closure. This is
because anonymous functions in javascript define their own this - and in
strict mode (which we are running in) that this is always undefined.

In ES6, the arrow function was introduced. One of the benefits of the
anonymous arrow function is that it does NOT define its own this...
precisely to allow a solution for the ugliness we have in the let ctrl =
this lines.

Update our http callback functions to be arrow functions and remove the
let ctrl = this lines.

Change-Id: I98a0011921e7c9473c05fa6091a213dac6fd92da