Split status and stream into typescript modules
This is the first in a set of patches aimed at stepping us to angular
(v2 -> v5 ... the thing that current dev is focused on) from angularjs
(v1 - the original version of the library) Angular uses typescript and
also is a bit more explicit about how things connect.

As a first step, introduce typescript support to the webpack config.
This will let us transition code piecemeal. The typescript support is
done as typescript -> es6 -> babel -> es5 rather than just typescript ->
es5. This is because babel supports adding polyfills into the output for
supporting older browsers, while typescript will compile to es5 but
needs polyfills to come from elsewhere. Since we're in a half and half
state to start with, it's easiest to just run the code through both.

Split status and stream into their own top-level module, which is
actually more in-line with how the output is organized. status and
stream are their own 'application', and have their own angularjs
'module'. In angular, the way a module attaches to the dom inverts, so
rather than having a module name on the body element, the initializer in
angular just attaches the module to the body.

Change-Id: Ida959da05df358994f4d11bb6f40f094d39a9541