Upgrade from angularjs (v1) to angular (v5)
Since we got started in all of this angular business, the angular folks
have released a major release of angular. The old v1 angular is known as
angularjs now, and starting at v2 the new codebase is just 'angular'.
While angularjs is still supported for now, angularjs vs. angular seems
to be more like zuulv2 vs. zuulv3 - the developers really want people to
be on the >=v2 series.

The notable differences are the angular is a bit more explicit/verbose,
and that it uses typescript instead of plain javscript. The increased
verbosity isn't the most popular with some fans of angularjs, but for
those of us who aren't breathing it every day I think the verbosity is
helpful.

The recommended code organization structure is adopted. Splitting main
into main+app.module+app.component is a bit wordy for now, but it sets
things up properly for us to use AOT on our production builds and to
initialize things properly for testing.

For zuul, there are notable changes to how the http client and location
service work, so the code related to those has been reworked.

$http has been reworked to use HttpClient - which defaults to grabbing
the remote json and which can do-so in a typesafe way.

$location has been reworked to use the routing module, which allows us
to pull both URL and Query String parameters in a structured manner. We
can similary pass query parameters to our output http requests.

Since routing is the new solution for $location, extract the navigation
bar into a re-usable component. The use of routing also allows us to not
depend on the webpack-level templating of appName to match page via
'ng-app="<%= htmlWebpackPlugin.options.app %>"' and instead use the
native angular structures.

Since we're shifting the files to typescript, update the webpack config
to use typescript as well - this helps catch some data-model syntax
errors.

The main.ejs file is a webpack template, not an angular template. Move
it to web/config with the other webpack files to make that clear.

Change-Id: Ida959da05df358994f4d11bb6f40f094d39a9541