Use glyphicons for status balls
This lets us remove the strange-looking 'import image into javascript'
(which totally works, but is slightly odd feeling).

It also lets us replace the images with font glyphs that we're loading
already for other things. And it provides a visual distinction between
the states that could be recognized by color-blind people.

Change-Id: Iacddfc5d75d49f55a3e4fd23a09a959cdccfeaec