Rename javascript package to zuul-dashboard
The python service is "zuul-web", but the javascript portion is
currently also being built as "zuul-web". While the javascript code is
intended to be served by zuul-web, not having its own name has made it
awkward to be clear about which codebase is being discussed.

Update the config to call the javascript code 'zuul-dashboard'.

Change-Id: I6e31ae23f74f887448f0a48147d91c156025e925