Perform late validation of secrets
This moves validation of the existence of secrets to after the
Job object has been constructed.  In the future, this will let
us re-validate Job objects if, for example, only the set of
secret objects has changed, without having to create new Job
objects.

We currently late-validate job parents; this is updated to use
the same new, more generic, mechanism as secrets.

This does incur a memory penalty in that we will be carrying around
start_mark objects which contain the original text of each
configuration entry attached to their resulting objects.  That will
increase the amount of memory used in the near term, but in the long
term, this will enable us to have fewer objects overall, once we
are able to re-use the same objects between different layouts.

The noop job needs a more correct fake playbook in order for it to
pass the new validation check.

The secret_not_found test is updated because the old version had
a configuration error (job not found) which, due to the later validation
of secrets, would have been propogated first.  This updates that
configuration so we proceed far enough to hit the secret not found
error.

Change-Id: I5bb673839ecf09939e4e57e7d548ef3fd96f767f