Update patch set 2
Patch Set 2:

> From an enterprise user point of view this is a problematic feature
 > for me as this circumvents the tenant concept of completely
 > separating projects. I would have to take extra care of specially
 > blocking this path.
 > 
I guess if you want to completely separate tenants, then you should already be restricting access to /{tenant}

 > What's your use case? I think getting all projects also can be
 > easily retrieved now in just a few calls to /tenant and loop over
 > the tenants with /{tenant}/projects.
The use case is to be able to issue an auto-hold request from a code-review comment where we only have the project name. To do that we need to know what are the tenants associated with a project. Indeed we could loop over /tenants, but this could be a simplified with the proposed endpoint.

Patch-set: 2
Reviewer: Gerrit User 9311 <9311@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0