Migrate status, stream and builds to angular
Update to use routing instead of $location. That also means extracting a
navigation component.

The earlier patches to split out top-level entrypoints for each page was
a mistake that winds up leading to a bunch of awkward code and
duplication. Re-integrate them into a single top-level app. Where before
we used a templated value for ng-app to decide which content was
associated with with page, now we just use the routing module for the
same, which lets us express url-structure into variables as well as pull
parameters.

Change-Id: Iaac42d628afb5304765242421aa5644bcd7aa529