Don't count non-live queue items in tenant list
The queue size in the tenant list should not include non-live
items. The initial thought was to exclude non-active items. However
non-active items in fact are queued items which are not in the active
window of a dependent pipeline. Thus they should be counted. The thing
we want to exclude is non-live items. Change that and add a test case
for that.

Change-Id: Ic63f0247fefca2cb3c8ea6c63eb405c4c4cf8dc0