Update patch set 7
Patch Set 7:

The last PS changes the patch in a way the errors are accumulated on the Tenant.layout object so that should fix the issue you mentioned for the dynamic reconfiguration. The errors accumulator array is now limited to the 10 first errors to save memory at scale. I also added the requested functional test to show that error detected during a dynamic reconfiguration are not kept on the tenant. Finally loading errors are only displayed in the log at the end of the loading or reloading tenant process as well as the name of tenant is display.

Patch-set: 7
Label: Verified=0