Fix logging in tests to be quiet when expected
During a normal test run, the intent is that the logging output should
not be displayed unless there is an error, but that if tests are run
outside of the context of a full run they will emit output.

The OS_LOG_DEFAULTS section was causing the logs to always spew. We're
not actually using it, so remove it. Also, update the logger setup to
set sane defaults for kazoo and gear.

Change-Id: I9a4e085c2cd7742d84ca0db67b5a6ece6d59eef2