Fix logging in tests to be quiet when expected
During a normal test run, the intent is that the logging output should
not be displayed unless there is an error, but that if tests are run
outside of the context of a full run they will emit output.

The OS_LOG_DEFAULTS section was not documented so mention it in the
testing docs. Also, update the logger setup to set sane defaults for
kazoo and gear.

Change-Id: I9a4e085c2cd7742d84ca0db67b5a6ece6d59eef2
Co-authored-by: Tobias Henkel <tobias.henkel@bmw.de>