Add license and downgrade exception to alembic template
Alembic uses a script template to generate new revisions. The generated
code is deficient in three ways that require copy-pasta on the part of the
developer. They are missing a license header. They have 'pass' in the
downgrade portion rather than throwing an exception about downgrades not
being supported. They are missing the table_prefix argument to upgrade.

Add all three to the template.

Change-Id: I7ff371706bd94dcd2744a32af13c74bc7cabdfdb