Update patch set 2
Patch Set 2:

Use case for a project being in more than one tenant - yes. Consider zuul-jobs. Or a project-config-like repo that is specific to the local installation but is needed by every tenant (like one containing a base job)

I think an option 4 might be to introduce auth and/or user roles. It seems like a server-side loop over tenants is a useful thing ... for an admin. That, of course, is a bigger deal that we should all discuss a lot before opening.

I'd lean towards option 1 for now while we discuss the various options.

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0