WIP Retry merge jobs
When gearman returns a WORKFAIL or WORKEXCEPTION it is not because
the operation (eg git merge) failed, but rather the worker failed
for some reason. This could include disconnects, full disks, or other
host errors. (A failed merge will return a WORKCOMPLETE with
appropriate data).

By allowing retries we make it easier to take workers (including
executors) offline as it does not matter if they stop abruptly.

Change-Id: Iaac4bedcf7303815d362b4315b11a813517888c1