Use typescript for webpack config
The next change updates the tree to use angular v5 which is typescript
based. As a baby step, update the webpack config from javascript to
typescript. Add a type annotations so that we can get compilation errors
from incorrect config. Also add a local workaround for an issue with the
upstream type definitions in the @types/webpack package.

Change-Id: I06b29f54f3bde4c52c2a9323baf9acff60a4efbc