Improve root api endpoint finding
The relative paths here weren't really doing their job. Instead, go
ahead and pull in the function we wrote for the later angular5 patch to
dynamically find the base href and use it to figure out what we should
be appending the api paths to.

Also, Remove double api prefix from tenants call

Erroneously added it in the previous patch while also updating
getSourceUrl to pre-pend the api string appropriately.

Change-Id: Ib58c4569dbe21824c138c57c6353454ea1e4bc71