Remove need for ZUUL_BASE_HREF build time option
Needing to specify ZUUL_BASE_HREF at build time for sub-url deployments
is makes things harder to use. We need to, however, be able to find the
root of the deployment so that routing will work.

Update the function that is already providing that information to infer
it from url structure. Leave ZUUL_BASE_HREF support there in case there
is anyone with a crazy deployment scheme that breaks the inferrance as
a get-out-of-jail card. (I can't come up with url schemes that break it,
but it's a no-op if not used, so the cost of keeping it is pretty low)

Change-Id: I103bba57cf1c09b7ed932ef0c27f485213e98f90