Update patch set 11
Patch Set 11: Code-Review+2

Change is fine but the vertical line isn't as neat as it should be due to the transparency behind the icons. We might need to add a class to the first and last changes in a queue to offset the line at each end.

Patch-set: 11
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2