Update patch set 3
Patch Set 3: Code-Review-2

This doesn't match the architectural model of Zuul.  Triggers respond to external stimuli and enqueue refs.  A "job" trigger doesn't make sense because it doesn't connect Zuul to a "job" (consider how existing triggers fit: Gerrit, GitHub, Timer, Zuul, as well as other likely ones: SMTP, IRC, MQTT, fedmesg).  This trigger doesn't enqueue a ref, instead it bypasses the project-pipeline config entirely.  Even if we adjust the architecture to support triggering an individual job instead of an entire job graph, it would need to be a subset of the job graph in a project's pipeline config.

Let's start with describing the use case and design something that fits.

Patch-set: 3
Label: Code-Review=-2