Update patch set 9
Patch Set 9: Code-Review-1

(1 comment)

This is looking really good now.  I left one question inline, and I have one general question:

After seeing this in practice, it can be confusing when the configuration is broken and someone proposes a .zuul.yaml change while unaware that the config is broken and they receive a message with the first configuration error (which has nothing to do with their change).  On the one hand, if someone were trying to fix the configuration, they would need that feedback.  However, if they weren't, then it's not clear what's going on.

Perhaps we should alter the message when reporting a dynamic configuration error when the running config is broken so that it says something like "Zuul's operating configuration has errors and can not be changed until they are corrected."

Finally, I think this warrants a release note since it's a significant user-visible behavior change.  Can you add a reno note for it?

Patch-set: 9
Label: Code-Review=-1