Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

I'm not seeing a lot of overlap myself, other than driver boilerplate.  I think it's okay to have two drivers.

-1: I think we want a release note to let people know there's a new driver.  Can you add a reno note?

I left a suggestion for improvement inline, but I don't think it's required in order to land this now.  But we might want it in the future.

Patch-set: 5