Update patch set 5
Patch Set 5: Code-Review-1

(5 comments)

This is a good start. I'm not super familiar with all the zuul internal interfaces, but I left some comments on the MQTT usage. There are somethings we'll need to add there before this is really read for use. The big ones are hierarchical topics and tls support.

Patch-set: 5
Label: Code-Review=-1