Update patch set 7
Patch Set 7: Code-Review-1

I like the abstraction, but I don't think we should put this in drivers.  I'd like to reserve drivers for interacting with other systems.  The rule of thumb might be: if the scheduler doesn't need to know about it, it shouldn't be in drivers.  (I think we should move bubblewrap out of drivers, but that can wait.)

Instead, let's just organize this under zuul.executor.sensors or something similar.

Patch-set: 7
Label: Code-Review=-1