Update patch set 2
Patch Set 2:

I think we need to at least toss a warning if not an error.

That said - if we're going to throw a warning or an error we have to detect the .yml files anyway, so we have to take on the cost associated with #2 mentioned by Jim.

I agree we don't need to necessarily copy all of Ansible's yaml choices - but part of our message is "if you're already using ansible for production, zuul is awesome for CI" - and since ansible tends to default to .yml for things I worry that we've got a large number of potential users primed for their early experience of us being us complaining about file extensions. I don't personally believe the value of having only one extension is offset by the potential for user pain or annoyance.

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0