Update patch set 9
Patch Set 9:

> (1 comment)
 > 
 > This is looking really good now.  I left one question inline, and I
 > have one general question:
 > 
 > After seeing this in practice, it can be confusing when the
 > configuration is broken and someone proposes a .zuul.yaml change
 > while unaware that the config is broken and they receive a message
 > with the first configuration error (which has nothing to do with
 > their change).  On the one hand, if someone were trying to fix the
 > configuration, they would need that feedback.  However, if they
 > weren't, then it's not clear what's going on.
 
Yes that's true the first error stacked was the one returned to the user when adding a .zuul.yaml and indeed usually not related to the user .zuul.yaml. The PS8 implements a solution to verify if the project:branch couple of the .zuul.yaml proposal is known to have a configuration issue thanks to the parse context that is now saved in the loading_errors stack (manager/__init__.py). If it is not the case then nothing is reported to the user and the change is processed a usual. 

Also, to return loading_errors to user, I attempted to implement a change in zuul web to expose issues in tenant/info endpoint: https://review.openstack.org/#/c/553873 (I need to refresh this patch).

And to fully expose via the web UI there is this one too: https://review.openstack.org/#/c/554652

 > Perhaps we should alter the message when reporting a dynamic
 > configuration error when the running config is broken so that it
 > says something like "Zuul's operating configuration has errors and
 > can not be changed until they are corrected."
 
As you prefer, either I keep the PS8 change or I set this kind a warning message.

 > Finally, I think this warrants a release note since it's a
 > significant user-visible behavior change.  Can you add a reno note
 > for it?

Yes I'm adding it.

Patch-set: 9