Localize build time in build overview
In the browser, we'd like to see timezone-adjusted, readable time
strings. We therefore convert the time we receive via the zuul api to
local time in the browser. However, This trequires us to store the
build time in UTC instead of local time. This might cause old build
times to be displayed with an incorrect offset.

Change-Id: Ie0cfce385854caa5adbd27f7f13042e7bfd41f1b