Update patch set 1
Patch Set 1: Code-Review+2

Not approving as I think it would be good to get some clarification on whether or not this migration should be updated or if we need a new migration entirely.  I think there are potential drawbacks to both approaches depending on whether or not you have deployed to production using a table prefix before this change got in then we attempt to modify the indexes in a future migration.

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2