WIP: Fix issue when using _blockPluginDirs
There seems to be an issue when we implictly add a project to the
ansible role path when it has a plugin. We block the plugin regardless
of the zuul run playbook actually using the project as a role.

This is to show the failure and brainstorm how to fix.

Change-Id: I93268434a9c37d974b7ec232db8ab18a5ef3a686
Signed-off-by: Paul Belanger <pabelanger@redhat.com>