Shift log streaming code into StreamComponent
Rather than an external function, make zuulStartStream a method on
StreamComponent. This puts us in a good position to refactor it to use
data binding.

The only functional changes made in this patch other than re-arranging
where the code is are making zuulStartStream not take this.route as a
parameter but instead use this.route which it has access to now.

Change-Id: I0da5ac3f4c345d49e0f385d7c8e89c78a976a9de