Update patch set 18
Patch Set 18:

We had a question in IRC that I'm curious about, before we merge this -- can you describe why this is helpful, versus just having the same script write output to disk and then triggering a reconfiguration?

If there's something about that process that makes this worthwhile, I'm happy to merge it.  But if it's not that big of a help, it might be nice to keep complexity in zuul and the number of config options to a minimum.

Patch-set: 18