Update patch set 18
Patch Set 18:

Sure, I think it simplify the process by not a having a configuration management system that run the script + trigger a reload. That's not a big difference but that can help to simplify Zuul integration with external systems that is good imo.

If in addition in the future we get a way to trigger a reload via a Zuul admin protected API then it makes the process really straightforward.

Also we can think to propose example scripts for external system such as for github where it reads the organization and compute the tenant configuration from repositories labels (zuul:config-repo, zuul:includes:jobs, ...) and have a tenant option to tell Zuul to reload when labels starting by 'zuul:' change.

Patch-set: 18
Reviewer: Gerrit User 6889 <6889@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0