Add role information to task in zuul_json callback
In comparison to other callback plugins like the default (stdout)
plugin, the role information for an executed task is missing in the
json output. To be complient to the provided task output containing
the name and uuid field, we've defined a similar data structure
containing the role information. This will only be added to the result
set if the task contained the necessary role information.

Change-Id: I8d94ba077e0bc90b5cf6510804bbd57c38184a9d