Use nodeenv for npm and yarn in tox
Add a wrapper installation script that will install npm and yarn into the
current virtualenv. If yarn is installed globally, the script is a
no-op.

If yarn is not installed globally, whenever tox thinks it needs to
create or re-create a virtualenv, install nodeenv then use it to
install yarn.

The javascript is already built/installed by setup.py install when yarn
is present, so when zuul is installed into the virtualenv the javascript
code should get built.

Change-Id: If360a3f0c6b3d74498f8aa063d8b1ae87daff101