Remove docker instructions and build:docker helper command
The docker instructions and build:docker helper command were put in
place when we first started working on this stack to try to reduce the
burden on people from the new toolchain. However, build:docker results
in root-owned files in zuul/web/static, and in general is complex and
fragile for not very much benefit now that nodeenv is in place for tox
runs.

The docker instructions in the docs, as well as the instructions for
people with old npm are similarly removed, as nodeenv handles these
cases as well.

In general, we should be encouraging people to get familiar with the
tools rather than hiding them.

Change-Id: I931d1101e025ba1236f4abc1d5e088ba5d113a5a