Add test for builds/ web route
This adds a simple test for the builds web route. Because this route
depends on the sql reporter functionality we update the sql-driver test
fixture config to accomodate the existing zuul web test expectations.
Namely that there is also a project1 project and that there are gate
jobs.

In the test we then confirm the builds/ url returns the expected count
of run jobs.

Change-Id: I51ff85a124fe75a7a0c4ec581f0b1f49c1aa0c05