Update patch set 4
Patch Set 4: Code-Review-1

I like this direction, but I think we need to go a bit further before we merge it.  We can either respin this, or add it as a followup to the end of the stack.  (I find it easier to review the finished product of docs changes anyway.)

The "install nodepool" and "install zuul" pages are so small and devoid of content, they look broken.  I think instead of splitting up the 'install' and 'setup' sections, we should lump adding the user, cloning, pip installing, making directories, and service unit files all together on those pages.  Don't start the services though.  That way we can have all the boring installation stuff in one place, then we can walk through writing config files and starting services in their own page.

Patch-set: 4
Label: Code-Review=-1