Update patch set 1
Patch Set 1:

In my opinion the buildset resources section of the spec is the most interesting. The first two sections (Containers That Behave Like a Machine & Native Container Workflow) are more about doing our traditional jobs (eg linting) in a container and provide alternative approaches in implementation.

Using nodepool to hold containers while zuul interacts with them seems cumbersome and hacky. However, it's perhaps a good solution to get us to containers without much effort. So if we wanted to reduce our resource usage by shifting to containers this may be a quick way of getting there. Otherwise I think we should focus on a more native container workflow.

If the job is building the container image and not doing it on a "full machine from nodepool" then from its current feature set I'm not sure what purpose Nodepool has? So rather than requiring nodepool to create an OpenShift project and set up the image repository etc, perhaps it is worth considering zuul or a new component to handle that. This will allow us to not arbitrarily couple nodepool's complexities with container concerns.

Back to buildsets. There should be a job that builds the images. This is because that's an important part of testing container workflow. Being able to then run tests against the image/container is also very useful. Finally, being able to use that image asset in some way is important. For example, a buildset might: 1) build and validate the image, 2) run smoke tests, then 3) upload a verified image to a public repository (ie, more like a release pipeline).

Having the job build the images means that we've lost some of the scale out gained from slave nodes. In other words, we're making the zuul-executors do all the work. This might be an argument for doing the building on a VM from nodepool, but that is also a lot of overhead for simpler images.

Additionally we need to consider how we do tests against a container? Are these simply ran by the job executor again?

Patch-set: 1