Update patch set 1
Patch Set 1:

I like the idea.  If this makes sense to you, I think I'd prefer putting the retry logic in the MergeClient.  At least to me, it makes sense to have the client retry the job, rather than have the job retry itself.  Obviously, the MergeJob would still be special in that it would have the count of attempts.  But that's just code organization, I think the implementation is sound.

I'm assuming this in WIP becaues it needs tests.

Patch-set: 1