Update patch set 2
Patch Set 2: Code-Review-1

I think once we add words explaining what this is all for, we'll want to keep it this way.  The idea was that we'd walk people through the process of setting up Zuul and explain it along the way.  That works better if we can make small steps and confirm that each one is working before going to the next.  We don't need this to start Zuul the first time, so let's leave it as an add-on.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1