Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

If we expect a large number of simultaneous gearman jobs running at the same time, then my comment is important.  I suspect that might be possible for the websocket address queries?  If not, then we could probably merge this as-is, mostly since we expect to stop using gearman for this sort of thing eventually.

Patch-set: 4
Label: Code-Review=-1