Update patch set 1
Patch Set 1:

(5 comments)

Joshua Hesketh wrote:

> In my opinion the buildset resources section of the spec is the most
> interesting. The first two sections (Containers That Behave Like a
> Machine & Native Container Workflow) are more about doing our
> traditional jobs (eg linting) in a container and provide alternative
> approaches in implementation.

The second, "Native Container Workflow" is definitely not intended to
be like our current jobs.  It's intended to support the workflow of
testing an application which runs in containers.

> Using nodepool to hold containers while zuul interacts with them
> seems cumbersome and hacky. However, it's perhaps a good solution to
> get us to containers without much effort. So if we wanted to reduce
> our resource usage by shifting to containers this may be a quick way
> of getting there. Otherwise I think we should focus on a more native
> container workflow.

This second pargraph makes me think you agree with what I just wrote,
but doesn't seem to agree with the first paragraph.  I'm slightly
confused.  :)

> If the job is building the container image and not doing it on a
> "full machine from nodepool" then from its current feature set I'm
> not sure what purpose Nodepool has? So rather than requiring
> nodepool to create an OpenShift project and set up the image
> repository etc, perhaps it is worth considering zuul or a new
> component to handle that. This will allow us to not arbitrarily
> couple nodepool's complexities with container concerns.

I think many of us have seen Nodepool as a "generic resource provider
for Zuul" for a while.  There's all kinds of useful things that Zuul
could request and Nodepool could provide, including more complex
scenarios like "two nodes on a private network".  We've tried to
restructure Nodepool so that we can accomodate extensions like this
in the future.

> Back to buildsets. There should be a job that builds the images.
> This is because that's an important part of testing container
> workflow. Being able to then run tests against the image/container
> is also very useful. Finally, being able to use that image asset in
> some way is important. For example, a buildset might: 1) build and
> validate the image, 2) run smoke tests, then 3) upload a verified
> image to a public repository (ie, more like a release pipeline).

Generally agreed -- but this does touch on an aspect we should probably
cover in the spec.  With Zuul, the product of a gate job is still an
interim result until it's merged.  So it would not be wise to upload
a gate image build to a public repo.  You could do what we do for
tarballs, and build the image again in post (trusting that you'll do
so with the same commit hashes).  Or you could upload the image to
an interim object store, and then promote it to public in a release
pipeline.  Either way, publishing a public image extends beyond the
gate buildset context.

> Having the job build the images means that we've lost some of the
> scale out gained from slave nodes. In other words, we're making the
> zuul-executors do all the work. This might be an argument for doing
> the building on a VM from nodepool, but that is also a lot of
> overhead for simpler images.

I wouldn't expect the executors to build the image -- if we're not
building the image on a vm, then I expect it to be built in a k8s
pod, and that shouldn't be running on the executor.

> Additionally we need to consider how we do tests against a
> container? Are these simply ran by the job executor again?

I'm starting to think I may have not conveyed something (that's why
it's a first draft!).  I expect we can write a job which uses the
k8s_raw ansible module to deploy k8s pods based on the images we
just built, and then runs tests (however they are implemented) against
them.  Those tests could be in their own pod, or they could be, say,
a python script running on a VM.  They can be implemented with anything
Zuul and Ansible can run.

Patch-set: 1